Add suppot for rated pools, refactor exchange.ts to make it modular
