+++ ref-finance/deployments.txt
@@ -72,6 +72,30 @@
         "timestamp": "1652092433193477496", May 9, 2022 - 1.4.5 
         "codeHash": "0x951d396c87f223a22b4e692082fb7d9597b6a5219ac2e41608aaa07697e5e5bc",
         "accountId": "v2.ref-finance.near"
+      },
+      {
+        "id": "8Ss6HCGqyBk4W9aWqHvLP5J4831u1yELh8MK34KgMEsB",
+        "receiptId": "8Ss6HCGqyBk4W9aWqHvLP5J4831u1yELh8MK34KgMEsB",
+        "blockNumber": "67601192",
+        "timestamp": "1655087786287951869", June 13, 2022 - 1.5.1
+        "codeHash": "0x77bbbfacbd5ecfdea17c05bf336f8f0b38e2d5c328545ec5ee36fd455c909f15",
+        "accountId": "v2.ref-finance.near"
+      },
+      {
+        "id": "ABcLrENE2eAhBpJ1EiKwN6Z2aFUwKBED9zoYCHCRmWyZ",
+        "receiptId": "ABcLrENE2eAhBpJ1EiKwN6Z2aFUwKBED9zoYCHCRmWyZ",
+        "blockNumber": "69864615",
+        "timestamp": "1657849890029170078", July 15, 2022 - 1.5.2
+        "codeHash": "0xd2b1e31e9a5d7690112a6dde782b731f8313ef963d1992c8e43c9cfc11ecec18",
+        "accountId": "v2.ref-finance.near"
+      },
+      {
+        "id": "uuVHXDMzux9fgekHGSV7YhRo5XQMZNzJaNyuQ3P9KiW",
+        "receiptId": "uuVHXDMzux9fgekHGSV7YhRo5XQMZNzJaNyuQ3P9KiW",
+        "blockNumber": "70754442",
+        "timestamp": "1658924287491898469", July 27, 2022 - 1.5.3
+        "codeHash": "0x6ac3fd69be9fd44f196d2505a477faa22b1326937bec07a5a4bdc99f91a45266",
+        "accountId": "v2.ref-finance.near"
       }
     ]
   }
